MobileCLDCode/02_Python/requirements.txt
pandas>=2.0
openpyxl>=3.1
PyYAML>=6.0
scikit-learn>=1.3
requests>=2.31
pdfplumber>=0.10
python-dateutil>=2.8
numpy>=1.24
